% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagginsclass.R
\name{BagginsClass}
\alias{BagginsClass}
\title{BagginsClass}
\value{
R6 object
}
\description{
BagginsClass：I am going on an adventure! this is an R6 object for rnaseq upstream workflow
}
\examples{
workflow <- BagginsClass$new(root_dir = ".")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{root_dir}}{character Path to the working directory for BAGGINS.}

\item{\code{bamfiles_dir}}{character Path to the directory containing aligned BAM files.}

\item{\code{countfiles_dir}}{character Path to the directory containing count files.}

\item{\code{refer_genome}}{character Path to the reference genome used for alignment.}

\item{\code{gtf_file}}{character Path to the annotation file in GTF format.}

\item{\code{samtools_exe}}{character Path to the SAMtools executable.}

\item{\code{htseq_exe}}{character Path to the HTSeq executable.}

\item{\code{star_exe}}{character Path to the STAR executable.}

\item{\code{clean_str}}{unnecessary str in the filenames}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rnaseq_workflow-new}{\code{BagginsClass$new()}}
\item \href{#method-rnaseq_workflow-info}{\code{BagginsClass$info()}}
\item \href{#method-rnaseq_workflow-initialize_dirs}{\code{BagginsClass$initialize_dirs()}}
\item \href{#method-rnaseq_workflow-fastq2matrix}{\code{BagginsClass$fastq2matrix()}}
\item \href{#method-rnaseq_workflow-clone}{\code{BagginsClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rnaseq_workflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-rnaseq_workflow-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagginsClass$new(
  root_dir = NA,
  bamfiles_dir = NULL,
  countfiles_dir = NULL,
  refer_genome = NULL,
  gtf_file = NULL,
  samtools_exe = NULL,
  htseq_exe = NULL,
  star_exe = NULL,
  clean_str = "Aligned.sortedByCoord.out.bam_counts.txt"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root_dir}}{根目录的路径}

\item{\code{bamfiles_dir}}{BAM文件目录的路径}

\item{\code{countfiles_dir}}{计数文件目录的路径}

\item{\code{refer_genome}}{参考基因组的路径}

\item{\code{gtf_file}}{GTF文件的路径}

\item{\code{samtools_exe}}{SAMtools可执行文件的路径}

\item{\code{htseq_exe}}{HTSeq可执行文件的路径}

\item{\code{star_exe}}{STAR可执行文件的路径}

\item{\code{clean_str}}{清理字符串的值}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rnaseq_workflow-info"></a>}}
\if{latex}{\out{\hypertarget{method-rnaseq_workflow-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagginsClass$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rnaseq_workflow-initialize_dirs"></a>}}
\if{latex}{\out{\hypertarget{method-rnaseq_workflow-initialize_dirs}{}}}
\subsection{Method \code{initialize_dirs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagginsClass$initialize_dirs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rnaseq_workflow-fastq2matrix"></a>}}
\if{latex}{\out{\hypertarget{method-rnaseq_workflow-fastq2matrix}{}}}
\subsection{Method \code{fastq2matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagginsClass$fastq2matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
转换后的矩阵
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rnaseq_workflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rnaseq_workflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagginsClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
