% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samwiseclass.R
\name{cpg_probe}
\alias{cpg_probe}
\alias{SamwiseClass}
\title{CpG island with gene annotation}
\format{
A data frame with CpG island with gene annotation.
SamwiseClass

SamwiseClass:It's the job that's never started takes longest to finish.this is an R6 object for rrbs upstream workflow
}
\value{
R6 object
}
\description{
A database of CpG island with gene annotation information.
}
\examples{
object2 <- SamwiseClass$new(root_dir = "batch1117", type = "cpg")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{root_dir}}{character NULL 根目录的路径。}

\item{\code{methyratio_dir}}{character NULL 存储methyratio文件的目录的路径。}

\item{\code{cpgmethy_dir}}{character NULL 存储CpG甲基化文件的目录的路径。}

\item{\code{cpgmethy_optimize}}{character NULL 优化后的CpG甲基化文件的目录的路径。}

\item{\code{promotermethy_optimize}}{character NULL 优化后的启动子甲基化文件的目录的路径。}

\item{\code{promotermethy_dir}}{character NULL 存储启动子甲基化文件的目录的路径。}

\item{\code{methyratio_py}}{character NULL 用于计算methyratio的Python脚本的路径。}

\item{\code{gene_fa}}{character NULL 基因组序列的FASTA文件的路径。}

\item{\code{bedtools_exe}}{character NULL BEDTools可执行文件的路径。}

\item{\code{cpg_bedfiles}}{character NULL CpG岛的bed文件的路径。}

\item{\code{promoter_bedfiles}}{character NULL 启动子区域的bed文件的路径。}

\item{\code{type}}{character "promoter" 甲基化分析的类型（"cpg"或"promoter"）。}

\item{\code{matrix_type}}{character "Beta-value" 矩阵表示的类型（"Beta-value"或"M-value"）。}

\item{\code{chrsex}}{character "dame" 染色体性别（"dame"或"male"）。}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rrbs_workflow-new}{\code{SamwiseClass$new()}}
\item \href{#method-rrbs_workflow-info}{\code{SamwiseClass$info()}}
\item \href{#method-rrbs_workflow-initialize_dirs}{\code{SamwiseClass$initialize_dirs()}}
\item \href{#method-rrbs_workflow-bam2methyratio}{\code{SamwiseClass$bam2methyratio()}}
\item \href{#method-rrbs_workflow-map2cpg4optimize}{\code{SamwiseClass$map2cpg4optimize()}}
\item \href{#method-rrbs_workflow-pack2matrix}{\code{SamwiseClass$pack2matrix()}}
\item \href{#method-rrbs_workflow-clone}{\code{SamwiseClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$new(
  root_dir = NA,
  methyratio_dir = NA,
  cpgmethy_dir = NA,
  cpgmethy_optimize = NA,
  promotermethy_optimize = NA,
  promotermethy_dir = NA,
  methyratio_py = "methratio.py",
  gene_fa = "Genedata/hg19.fa",
  bedtools_exe = "bedtools",
  cpg_bedfiles = "Genedata/cpg_islands_hg19.bedgraph",
  promoter_bedfiles = "Genedata/promoter_hg19_ensGene.bedgraph",
  type = "cpg",
  matrix_type = "Beta-value",
  chrsex = "dame"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root_dir}}{character NULL 根目录的路径。}

\item{\code{methyratio_dir}}{character NULL 存储methyratio文件的目录的路径。}

\item{\code{cpgmethy_dir}}{character NULL 存储CpG甲基化文件的目录的路径。}

\item{\code{cpgmethy_optimize}}{character NULL 优化后的CpG甲基化文件的目录的路径。}

\item{\code{promotermethy_optimize}}{character NULL 优化后的启动子甲基化文件的目录的路径。}

\item{\code{promotermethy_dir}}{character NULL 存储启动子甲基化文件的目录的路径。}

\item{\code{methyratio_py}}{character NULL 用于计算methyratio的Python脚本的路径。}

\item{\code{gene_fa}}{character NULL 基因组序列的FASTA文件的路径。}

\item{\code{bedtools_exe}}{character NULL BEDTools可执行文件的路径。}

\item{\code{cpg_bedfiles}}{character NULL CpG岛的bed文件的路径。}

\item{\code{promoter_bedfiles}}{character NULL 启动子区域的bed文件的路径。}

\item{\code{type}}{character "promoter" 甲基化分析的类型（"cpg"或"promoter"）。}

\item{\code{matrix_type}}{character "Beta-value" 矩阵表示的类型（"Beta-value"或"M-value"）。}

\item{\code{chrsex}}{character "dame" 染色体性别（"dame"或"male"）。}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
R6对象的实例。
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-info"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
str
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-initialize_dirs"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-initialize_dirs}{}}}
\subsection{Method \code{initialize_dirs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$initialize_dirs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-bam2methyratio"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-bam2methyratio}{}}}
\subsection{Method \code{bam2methyratio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$bam2methyratio()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
methyratio
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-map2cpg4optimize"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-map2cpg4optimize}{}}}
\subsection{Method \code{map2cpg4optimize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$map2cpg4optimize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
优化后的bedgraph文件。
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-pack2matrix"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-pack2matrix}{}}}
\subsection{Method \code{pack2matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$pack2matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
矩阵格式的数据框。
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rrbs_workflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rrbs_workflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamwiseClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
