% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entclass.R
\name{EntClass}
\alias{EntClass}
\title{EntClass}
\value{
R6 object
}
\description{
EntClass: We are not hasty. What we do, we do after long consideration and mostly we do not do anything.
this is an R6 object for general array/matrix
}
\examples{
Entitem <- EntClass$new(array = NULL,pdata = NULL)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array}}{data.frame data.frame of expression/methylation matrix data, first column is the probel id}

\item{\code{pdata}}{data.frame data.frame of the clinical data of the related matrix data}

\item{\code{probel}}{vector the probel of the matrix}

\item{\code{group_index}}{char the colname of the group infomation in pdata}

\item{\code{batch_index}}{char the colname of the batch_index infomation in pdata}

\item{\code{barcode}}{char the colname of the barcode infomation in pdata}

\item{\code{pca}}{data.frame the PCA analysis result}

\item{\code{tsne}}{data.frame the T-SNE analysis result}

\item{\code{normal_array}}{data.frame the normalized array}

\item{\code{DA_result}}{data.frame the differential analysis result}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EntClass/MatrixSet-new}{\code{EntClass$new()}}
\item \href{#method-EntClass/MatrixSet-info}{\code{EntClass$info()}}
\item \href{#method-EntClass/MatrixSet-feature_filter}{\code{EntClass$feature_filter()}}
\item \href{#method-EntClass/MatrixSet-PlotPCA}{\code{EntClass$PlotPCA()}}
\item \href{#method-EntClass/MatrixSet-PlotSNE}{\code{EntClass$PlotSNE()}}
\item \href{#method-EntClass/MatrixSet-toExprSet}{\code{EntClass$toExprSet()}}
\item \href{#method-EntClass/MatrixSet-toMethySet}{\code{EntClass$toMethySet()}}
\item \href{#method-EntClass/MatrixSet-normalization_transform}{\code{EntClass$normalization_transform()}}
\item \href{#method-EntClass/MatrixSet-DifferentAnalysis_workflow}{\code{EntClass$DifferentAnalysis_workflow()}}
\item \href{#method-EntClass/MatrixSet-clone}{\code{EntClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$new(
  array = NULL,
  pdata = NULL,
  group_index = "group",
  batch_index = "batch",
  barcode_index = "sampleid"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array}}{data.frame data.frame of expression/methylation matrix data, first column is the probel id}

\item{\code{pdata}}{data.frame data.frame of the clinical data of the related matrix data}

\item{\code{group_index}}{char the colname of the group infomation in pdata}

\item{\code{batch_index}}{char the colname of the batch_index infomation in pdata}

\item{\code{barcode_index}}{char the colname of the barcode infomation in pdata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-info"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
str
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-feature_filter"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-feature_filter}{}}}
\subsection{Method \code{feature_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$feature_filter(pct = 0.8)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pct}}{num range(0-1), the ratetio to filter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-PlotPCA"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-PlotPCA}{}}}
\subsection{Method \code{PlotPCA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$PlotPCA(ntop = 500)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ntop}}{num select the ntop rows by variance, default as 500}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-PlotSNE"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-PlotSNE}{}}}
\subsection{Method \code{PlotSNE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$PlotSNE(ntop = 500, dim = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ntop}}{num select the ntop rows by variance, default as 500}

\item{\code{dim}}{num the dim of Rtsne,default as 2}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-toExprSet"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-toExprSet}{}}}
\subsection{Method \code{toExprSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$toExprSet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
R6_object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-toMethySet"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-toMethySet}{}}}
\subsection{Method \code{toMethySet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$toMethySet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
R6_object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-normalization_transform"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-normalization_transform}{}}}
\subsection{Method \code{normalization_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$normalization_transform(.keep = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.keep}}{logit whether to replace the old array directly,default as TRUE (create the normal_array)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-DifferentAnalysis_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-DifferentAnalysis_workflow}{}}}
\subsection{Method \code{DifferentAnalysis_workflow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$DifferentAnalysis_workflow(adjust.method = "BH", adjPVal = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adjust.method}}{chr the adjust.method of limma}

\item{\code{adjPVal}}{num the adjust p value, default as 0.05}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntClass/MatrixSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EntClass/MatrixSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
