% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprentclass.R
\name{ExprEntClass}
\alias{ExprEntClass}
\title{ExprEntClass}
\value{
R6 object
}
\description{
ExprEntClass: We are not orcs, to be smashed and driven. We are Ents, and we have our own ways.
this is an R6 object for RNA expression array/matrix analysis,herited from EntClass.
}
\examples{
ExprEntitem <- ExprEntClass$new()
}
\section{Super class}{
\code{erebor::EntClass/MatrixSet} -> \code{ExprEntClass/ExprSet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hetmap}}{the hetmap}

\item{\code{vocalno}}{the vocalno graph}

\item{\code{gsva}}{matrix the gsva analysis result}

\item{\code{enrichment}}{list the enrichment analysis result}

\item{\code{removebatch_array}}{data.frame the array with batch effects removed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExprEntClass/ExprSet-new}{\code{ExprEntClass$new()}}
\item \href{#method-ExprEntClass/ExprSet-info}{\code{ExprEntClass$info()}}
\item \href{#method-ExprEntClass/ExprSet-removeBatch}{\code{ExprEntClass$removeBatch()}}
\item \href{#method-ExprEntClass/ExprSet-Plot_hetmap}{\code{ExprEntClass$Plot_hetmap()}}
\item \href{#method-ExprEntClass/ExprSet-Plot_vocalno}{\code{ExprEntClass$Plot_vocalno()}}
\item \href{#method-ExprEntClass/ExprSet-probel_annotate}{\code{ExprEntClass$probel_annotate()}}
\item \href{#method-ExprEntClass/ExprSet-gsva_enrich}{\code{ExprEntClass$gsva_enrich()}}
\item \href{#method-ExprEntClass/ExprSet-gene_enrich}{\code{ExprEntClass$gene_enrich()}}
\item \href{#method-ExprEntClass/ExprSet-clone}{\code{ExprEntClass$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="DifferentAnalysis_workflow"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-DifferentAnalysis_workflow'><code>erebor::EntClass/MatrixSet$DifferentAnalysis_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="PlotPCA"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-PlotPCA'><code>erebor::EntClass/MatrixSet$PlotPCA()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="PlotSNE"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-PlotSNE'><code>erebor::EntClass/MatrixSet$PlotSNE()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="feature_filter"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-feature_filter'><code>erebor::EntClass/MatrixSet$feature_filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="normalization_transform"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-normalization_transform'><code>erebor::EntClass/MatrixSet$normalization_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="toExprSet"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-toExprSet'><code>erebor::EntClass/MatrixSet$toExprSet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="toMethySet"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-toMethySet'><code>erebor::EntClass/MatrixSet$toMethySet()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$new(removebatch_array = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{removebatch_array}}{logit whether to remove the batch effect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-info"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
str
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-removeBatch"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-removeBatch}{}}}
\subsection{Method \code{removeBatch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$removeBatch(use_normalize = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_normalize}}{logit use normalize data when true}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-Plot_hetmap"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-Plot_hetmap}{}}}
\subsection{Method \code{Plot_hetmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$Plot_hetmap(p_thred = 0.05, logFC_thred = 1, topn = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p_thred}}{default as 0.05}

\item{\code{logFC_thred}}{default as 1}

\item{\code{topn}}{default as 30}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
fig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-Plot_vocalno"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-Plot_vocalno}{}}}
\subsection{Method \code{Plot_vocalno()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$Plot_vocalno(style_id, logFC_thred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style_id}}{default as 1,2}

\item{\code{logFC_thred}}{default as 1}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
fig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-probel_annotate"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-probel_annotate}{}}}
\subsection{Method \code{probel_annotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$probel_annotate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-gsva_enrich"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-gsva_enrich}{}}}
\subsection{Method \code{gsva_enrich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$gsva_enrich(geneSets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneSets}}{list the geneSets list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-gene_enrich"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-gene_enrich}{}}}
\subsection{Method \code{gene_enrich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$gene_enrich(pathwaydb, genedb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathwaydb}}{chr go or kegg}

\item{\code{genedb}}{data.frame the genedb data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExprEntClass/ExprSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExprEntClass/ExprSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExprEntClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
