% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylentclass.R
\name{MethylEntClass}
\alias{MethylEntClass}
\title{MethylEntClass}
\value{
R6_object
}
\description{
MethylEntClass : The world is changing. I feel it in the water. I feel it in the earth. I smell it in the air. Much that once was is now lost, for none now live who remember it.
}
\examples{
MethylEnt <- MethylEntClass$new()
}
\section{Super class}{
\code{erebor::EntClass/MatrixSet} -> \code{MethylEntClass/MethySet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MethylEntClass/MethySet-Plot_hetmap}{\code{MethylEntClass$Plot_hetmap()}}
\item \href{#method-MethylEntClass/MethySet-Plot_vocalno}{\code{MethylEntClass$Plot_vocalno()}}
\item \href{#method-MethylEntClass/MethySet-matrix_annotate}{\code{MethylEntClass$matrix_annotate()}}
\item \href{#method-MethylEntClass/MethySet-gene_enrich}{\code{MethylEntClass$gene_enrich()}}
\item \href{#method-MethylEntClass/MethySet-Mvalue_transform}{\code{MethylEntClass$Mvalue_transform()}}
\item \href{#method-MethylEntClass/MethySet-clone}{\code{MethylEntClass$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="DifferentAnalysis_workflow"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-DifferentAnalysis_workflow'><code>erebor::EntClass/MatrixSet$DifferentAnalysis_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="PlotPCA"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-PlotPCA'><code>erebor::EntClass/MatrixSet$PlotPCA()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="PlotSNE"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-PlotSNE'><code>erebor::EntClass/MatrixSet$PlotSNE()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="feature_filter"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-feature_filter'><code>erebor::EntClass/MatrixSet$feature_filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="info"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-info'><code>erebor::EntClass/MatrixSet$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="initialize"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-initialize'><code>erebor::EntClass/MatrixSet$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="normalization_transform"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-normalization_transform'><code>erebor::EntClass/MatrixSet$normalization_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="probel_annotate"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-probel_annotate'><code>erebor::EntClass/MatrixSet$probel_annotate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="removeBatch"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-removeBatch'><code>erebor::EntClass/MatrixSet$removeBatch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="toExprSet"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-toExprSet'><code>erebor::EntClass/MatrixSet$toExprSet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="toMethySet"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-toMethySet'><code>erebor::EntClass/MatrixSet$toMethySet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="erebor" data-topic="EntClass/MatrixSet" data-id="use_removebatch"><a href='../../erebor/html/EntClass/MatrixSet.html#method-EntClass/MatrixSet-use_removebatch'><code>erebor::EntClass/MatrixSet$use_removebatch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylEntClass/MethySet-Plot_hetmap"></a>}}
\if{latex}{\out{\hypertarget{method-MethylEntClass/MethySet-Plot_hetmap}{}}}
\subsection{Method \code{Plot_hetmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylEntClass$Plot_hetmap(p_thred = 0.05, logFC_thred = 1, topn = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p_thred}}{default as 0.05}

\item{\code{logFC_thred}}{default as 1}

\item{\code{topn}}{default as 30}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
fig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylEntClass/MethySet-Plot_vocalno"></a>}}
\if{latex}{\out{\hypertarget{method-MethylEntClass/MethySet-Plot_vocalno}{}}}
\subsection{Method \code{Plot_vocalno()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylEntClass$Plot_vocalno(style_id, logFC_thred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style_id}}{default as 1,2}

\item{\code{logFC_thred}}{default as 1}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
fig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylEntClass/MethySet-matrix_annotate"></a>}}
\if{latex}{\out{\hypertarget{method-MethylEntClass/MethySet-matrix_annotate}{}}}
\subsection{Method \code{matrix_annotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylEntClass$matrix_annotate(
  df_object = c("DA_result", "array"),
  type = "cpg"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_object}}{the converted data}

\item{\code{type}}{chr promoter or cpg}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylEntClass/MethySet-gene_enrich"></a>}}
\if{latex}{\out{\hypertarget{method-MethylEntClass/MethySet-gene_enrich}{}}}
\subsection{Method \code{gene_enrich()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylEntClass$gene_enrich(pathwaydb, genedb = Entrez_Gene_Id_db)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathwaydb}}{chr go or kegg}

\item{\code{genedb}}{data.frame the genedb data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylEntClass/MethySet-Mvalue_transform"></a>}}
\if{latex}{\out{\hypertarget{method-MethylEntClass/MethySet-Mvalue_transform}{}}}
\subsection{Method \code{Mvalue_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylEntClass$Mvalue_transform()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MethylEntClass/MethySet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MethylEntClass/MethySet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethylEntClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
